modulo/requirements.txt
requests
python-dotenv
redis
schedule
flask